add the ability to save sample rate and channels to csv header
This commit adds a new function called save_multi_csv() which allows you
to write multiple channels to a csv file as well as store the sample
rate in the header file. These can then be loaded by the Keysight AWG
hardware which will automatically update the sample rate and load the
sequences to the right channels.